fix: cannot get cart

+++ backend/src/cart/cart.service.ts
@@ -20,18 +20,13 @@ export class CartService {
     return {
       id: cart.id,
       userId: cart.userId,
-      items: cart.items.map((item: any) =>
-        this.mapCartItemToResponse(item, vouchers),
-      ),
-      createdAt: cart.createdAt,
-      updatedAt: cart?.updatedAt,
+      items: cart.items.map((item: any) => this.mapCartItemToResponse(item, vouchers)),
+      createdAt: cart.createdAt.toISOString(),
+      updatedAt: cart.updatedAt ? cart.updatedAt.toISOString() : null,
     };
   }
 
-  private mapCartItemToResponse(
-    cartItem: any,
-    vouchers: any[] = [],
-  ): CartItemResponseDto {
+  private mapCartItemToResponse(cartItem: any, vouchers: any[] = []): CartItemResponseDto {
     const relatedVouchers = vouchers.filter((v) =>
       v.conditions.productIds.includes(cartItem.variant.product.id),
     );
@@ -65,13 +60,30 @@ export class CartService {
               id: cartItem.variant.product.store.id,
               name: cartItem.variant.product.store.name,
             }
-          : undefined,
+          : undefined
       },
     };
   }
 
   async getCartByUserId(userId: string): Promise<CartResponseDto> {
-    const cart = await this.validateCart(userId);
+    const cart = await this.prisma.cart.findUnique({
+      where: { userId },
+      include: {
+        items: {
+          include: {
+            variant: {
+              include: {
+                product: {
+                  include: {
+                    store: true,
+                  },
+                },
+              },
+            }
+          },
+        },
+      },
+    });
 
     if (!cart) {
       const newCart = await this.prisma.cart.create({
@@ -81,7 +93,7 @@ export class CartService {
         include: {
           items: {
             include: {
-              variant: true,
+              variant: true
             },
           },
         },
@@ -95,6 +107,7 @@ export class CartService {
         isActive: true,
         startDate: { lte: currentDate },
         endDate: { gte: currentDate },
+        applicableFor: "SPECIFIC_PRODUCTS",
       },
     });
 
@@ -106,6 +119,7 @@ export class CartService {
     data: AddCartItemDto,
   ): Promise<CartItemResponseDto> {
     return this.prisma.$transaction(async (prisma) => {
+      
       const productVariant = await prisma.productVariant.findUnique({
         where: { id: data.variantId },
         include: {
@@ -117,12 +131,14 @@ export class CartService {
         throw new NotFoundException('Biến thể sản phẩm không tồn tại');
       }
 
-      if (!productVariant.product) {
+      if (!productVariant.product){
         throw new NotFoundException('Sản phẩm không tồn tại');
       }
 
       if (productVariant.quantity < data.quantity) {
-        throw new BadRequestException(`Số lượng sản phẩm còn lại không đủ!`);
+        throw new BadRequestException(
+          `Số lượng sản phẩm còn lại không đủ!`,
+        );
       }
 
       let cart = await prisma.cart.findUnique({
@@ -140,7 +156,7 @@ export class CartService {
       const existingItem = await prisma.cartItem.findFirst({
         where: {
           cartId: cart.id,
-          variantId: data.variantId,
+          variantId: data.variantId
         },
       });
 
@@ -159,7 +175,8 @@ export class CartService {
           },
         });
         return this.mapCartItemToResponse(updatedItem);
-      } else {
+      } 
+      else {
         const newItem = await prisma.cartItem.create({
           data: {
             cartId: cart.id,
@@ -167,13 +184,13 @@ export class CartService {
             quantity: data.quantity,
           },
           include: {
-            variant: {
+           variant: {
               include: {
                 product: {
                   include: {
                     store: true,
                   },
-                },
+                }
               },
             },
           },
@@ -188,13 +205,26 @@ export class CartService {
     data: UpdateCartItemDto,
   ): Promise<CartItemResponseDto> {
     return this.prisma.$transaction(async (prisma) => {
-      const cart = await this.validateCart(userId);
+      const cart = await prisma.cart.findUnique({
+        where: { userId },
+        include: {
+          items: {
+            include: {
+              variant: {
+                include: {
+                  product: true,
+                },
+              },
+            },
+          }
+        },
+      });
 
       if (!cart) {
         throw new NotFoundException('Giỏ hàng không tồn tại');
       }
 
-      const cartItem = cart?.items.find((item) => item.id === data.cartItemId);
+      const cartItem = cart?.items.find(item => item.id === data.cartItemId);
 
       if (!cartItem) {
         throw new NotFoundException('Item trong giỏ hàng không tồn tại');
@@ -204,12 +234,14 @@ export class CartService {
         throw new NotFoundException('Biến thể sản phẩm không tồn tại');
       }
 
-      if (!cartItem.product) {
+      if (!cartItem.variant.product) {
         throw new NotFoundException('Sản phẩm không tồn tại');
       }
 
       if (data.quantity > cartItem.variant.quantity) {
-        throw new BadRequestException(`Số lượng sản phẩm còn lại không đủ!`);
+        throw new BadRequestException(
+          `Số lượng sản phẩm còn lại không đủ!`,
+        );
       }
       const updatedItem = await prisma.cartItem.update({
         where: { id: data.cartItemId },
@@ -225,7 +257,7 @@ export class CartService {
                 },
               },
             },
-          },
+          }
         },
       });
       return this.mapCartItemToResponse(updatedItem);
@@ -236,14 +268,20 @@ export class CartService {
     userId: string,
     cartItemId: string,
   ): Promise<RemoveCartItemResponseDto> {
-    const cart = await this.validateCart(userId);
+    const cart = await this.prisma.cart.findUnique({
+      where: { userId },
+      include: {
+        items: true,
+      },
+    });
 
     if (!cart) {
-      throw new NotFoundException('Giỏ hàng không tồn tại');
+      throw new NotFoundException('Cart not found');
     }
 
-    if (!cart.items.some((item) => item.id === cartItemId)) {
-      throw new NotFoundException('Sản phẩm không có trong giỏ hàng');
+    const cartItem = cart.items.find((item) => item.id === cartItemId);
+    if (!cartItem) {
+      throw new NotFoundException('Cart item not found');
     }
 
     const deletedItem = await this.prisma.cartItem.delete({
@@ -257,7 +295,13 @@ export class CartService {
   }
 
   async clearCart(userId: string): Promise<CartResponseDto> {
-    const cart = await this.validateCart(userId);
+    const cart = await this.prisma.cart.findUnique({
+      where: { userId },
+    });
+
+    if (!cart) {
+      throw new NotFoundException('Cart not found');
+    }
 
     await this.prisma.cartItem.deleteMany({
       where: { cartId: cart.id },
@@ -279,32 +323,5 @@ export class CartService {
     });
 
     return this.mapCartToResponse(updatedCart);
-  }
-
-  async validateCart(userId: string): Promise<CartResponseDto> {
-    const cart = await this.prisma.cart.findUnique({
-      where: { userId },
-      include: {
-        items: {
-          include: {
-            variant: {
-              include: {
-                product: {
-                  include: {
-                    store: true,
-                  },
-                },
-              },
-            },
-          },
-        },
-      },
-    });
-
-    if (!cart) {
-      throw new NotFoundException('Giỏ hàng không tồn tại');
-    }
-
-    return this.mapCartToResponse(cart);
   }
 }
